apps/core-engine/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
pyjwt>=2.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
